Cast allocation results in dsv28dec.h.
While assigning a `void*`` to another pointer type is valid in C, it is invalid in C++ (not just a warning that can be disabled).

To allow dsv28dec.h to be included in C++ files (regardless of using `extern "C"``), these allocation results are now manually cast to the correct types.

To test this, simply run the compile command in C++ mode, which fails before this patch and works after:

`cc -O3 -o dsv28dec -x c++  d28_dec_main.c`